+++ src/main/kotlin/com/interrupt/server/career/dto/CareerJobDto.kt
@@ -1,7 +1,7 @@
 package com.interrupt.server.career.dto
 
 import com.interrupt.server.career.entity.CareerJob
-import com.interrupt.server.career.job.dto.JobDto
+import com.interrupt.server.job.dto.JobDto
 
 data class CareerJobDto(
     val career: CareerDto,
