+++ src/main/kotlin/com/interrupt/server/career/entity/CareerJob.kt
@@ -1,6 +1,6 @@
 package com.interrupt.server.career.entity
 
-import com.interrupt.server.career.job.entity.Job
+import com.interrupt.server.job.entity.Job
 import com.interrupt.server.common.entity.SoftDeleteBaseEntity
 import jakarta.persistence.Entity
 import jakarta.persistence.JoinColumn
