+++ src/main/kotlin/com/interrupt/server/job/entity/Job.kt
@@ -1,18 +1,18 @@
-package com.interrupt.server.career.skill.entity
+package com.interrupt.server.job.entity
 
 import com.interrupt.server.common.entity.SoftDeleteBaseEntity
 import jakarta.persistence.*
 
 @Entity
 @Table(
-    name = "skill",
-    uniqueConstraints = [UniqueConstraint(name = "uk_skill_name", columnNames = ["name"])]
+    name = "job",
+    uniqueConstraints = [UniqueConstraint(name = "uk_job_name", columnNames = ["name"])]
 )
-class Skill(
+class Job(
     @field:Column(name = "name", nullable = false, unique = true)
     var name: String,
     @field:ManyToOne(fetch = FetchType.LAZY)
-    @field:JoinColumn(name = "skill_group_id", nullable = false)
-    var skillGroup: SkillGroup
+    @field:JoinColumn(name = "job_group_id", nullable = false)
+    var jobGroup: JobGroup
 ): SoftDeleteBaseEntity() {
 }
