+++ src/main/kotlin/com/interrupt/server/job/repository/CareerJobQueryRepository.kt
@@ -1,9 +1,9 @@
-package com.interrupt.server.career.job.repository
+package com.interrupt.server.job.repository
 
-import com.interrupt.server.career.job.dto.JobDto
-import com.interrupt.server.career.job.dto.JobGroupDto
-import com.interrupt.server.career.job.entity.Job
-import com.interrupt.server.career.job.entity.JobGroup
+import com.interrupt.server.job.dto.JobDto
+import com.interrupt.server.job.dto.JobGroupDto
+import com.interrupt.server.job.entity.Job
+import com.interrupt.server.job.entity.JobGroup
 import org.springframework.stereotype.Repository
 
 @Repository
