+++ src/main/kotlin/com/interrupt/server/job/repository/CareerJobRepository.kt
@@ -1,6 +1,6 @@
-package com.interrupt.server.career.job.repository
+package com.interrupt.server.job.repository
 
-import com.interrupt.server.career.job.entity.Job
+import com.interrupt.server.job.entity.Job
 import com.linecorp.kotlinjdsl.support.spring.data.jpa.repository.KotlinJdslJpqlExecutor
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
