+++ src/main/kotlin/com/interrupt/server/skill/dto/SkillDto.kt
@@ -1,6 +1,6 @@
-package com.interrupt.server.career.skill.dto
+package com.interrupt.server.skill.dto
 
-import com.interrupt.server.career.skill.entity.Skill
+import com.interrupt.server.skill.entity.Skill
 import com.interrupt.server.common.annotation.RedisEntity
 
 @RedisEntity
@@ -10,6 +10,6 @@ data class SkillDto(
     val skillGroup: SkillGroupDto
 ) {
 
-    fun toEntity(): Skill  = Skill(name, skillGroup.toEntity())
+    fun toEntity(): Skill = Skill(name, skillGroup.toEntity())
 
 }
