+++ src/main/kotlin/com/interrupt/server/skill/repository/CareerSkillRepository.kt
@@ -1,6 +1,6 @@
-package com.interrupt.server.career.skill.repository
+package com.interrupt.server.skill.repository
 
-import com.interrupt.server.career.skill.entity.Skill
+import com.interrupt.server.skill.entity.Skill
 import com.linecorp.kotlinjdsl.support.spring.data.jpa.repository.KotlinJdslJpqlExecutor
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
